Add agent management functionality and database seeding script
Introduce agent CRUD operations, API endpoints, and a script to populate the database with agent data.

+++ ui/server/db/models/agent.py
@@ -12,10 +12,10 @@ class Agent(Base):
     skills = Column(ARRAY(Text), nullable=False)
     base_price = Column(Float, nullable=False)
     dynamic_price = Column(Float, nullable=False)
-    load = Column(Integer, nullable=False, default=0)
+    load = Column(Float, nullable=False, default=0.0)
     rating = Column(Float, nullable=False, default=5.0)
     jobs_completed = Column(Integer, nullable=False, default=0)
     endpoint_url = Column(Text, nullable=False)
     capabilities = Column(ARRAY(Text), nullable=False)
-    avg_response_time = Column(Integer, nullable=False, default=1000)  # ms
+    avg_response_time = Column(Float, nullable=False, default=1000.0)  # ms
     availability = Column(Boolean, nullable=False, default=True)
